// EmployeeManager.ClientApp/src/app/core/api/generated/model/employeeDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Gender } from './gender';


export interface EmployeeDto { 
    id?: number;
    firstName?: string | null;
    lastName?: string | null;
    middleName?: string | null;
    birthDate?: string;
    gender?: Gender;
    email?: string | null;
    phoneNumber?: string | null;
    joinedDate?: string;
    exitedDate?: string | null;
    superiorId?: number | null;
    addressId?: number | null;
    jobCategoryIds?: Array<number> | null;
    salaryIds?: Array<number> | null;
    subordinateIds?: Array<number> | null;
}
export namespace EmployeeDto {
}