// EmployeeManager.ClientApp/src/app/core/api/generated/model/employeeUpdateDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Gender } from './gender';


export interface EmployeeUpdateDto { 
    firstName?: string | null;
    lastName?: string | null;
    middleName?: string | null;
    birthDate?: string;
    gender?: Gender;
    email?: string | null;
    phoneNumber?: string | null;
    joinedDate?: string;
    exitedDate?: string | null;
    superiorId?: number | null;
    addressId?: number | null;
    jobCategoryIds?: Array<number> | null;
}
export namespace EmployeeUpdateDto {
}