// EmployeeManager.ClientApp/src/app/core/api/generated/model/cityDto.ts
/**
 * EmployeeManager.WebAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CityDto { 
    id?: number;
    name?: string | null;
    countryId?: number | null;
}